basic seperation to modules, group support
