+++ utils.py
@@ -1,0 +1,9 @@
+# Return chat ID for an Update object
+def get_chat_id(update=None):
+    if update:
+        if update.message:
+            return update.message.chat_id
+        elif update.callback_query:
+            return update.callback_query.from_user["id"]
+    else:
+        return None
